Agent: Rename zerologon_utils.vuln_assessment to dc_utils
