+++ monkey/infection_monkey/exploit/zerologon.py
@@ -25,9 +25,9 @@ from common.tags import (
 )
 from infection_monkey.exploit.HostExploiter import HostExploiter
 from infection_monkey.exploit.tools.wmi_tools import WmiTools
+from infection_monkey.exploit.zerologon_utils.dc_utils import connect_to_dc, get_dc_details
 from infection_monkey.exploit.zerologon_utils.dump_secrets import DumpSecrets
 from infection_monkey.exploit.zerologon_utils.options import OptionsForSecretsdump
-from infection_monkey.exploit.zerologon_utils.vuln_assessment import connect_to_dc, get_dc_details
 from infection_monkey.exploit.zerologon_utils.wmiexec import Wmiexec
 from infection_monkey.i_puppet import ExploiterResultData
 from infection_monkey.utils.capture_output import StdoutCapture
