+++ monkey/tests/unit_tests/infection_monkey/exploit/zerologon_utils/test_dc_utils.py
@@ -21,8 +21,8 @@ def _get_stub_queryIPForName(netbios_names):
 
 
 @pytest.mark.slow
-def test_get_dc_details_multiple_netbios_names(host, monkeypatch):
-    from infection_monkey.exploit.zerologon_utils.vuln_assessment import get_dc_details
+def test_dc_utils__get_dc_details_multiple_netbios_names(host, monkeypatch):
+    from infection_monkey.exploit.zerologon_utils.dc_utils import get_dc_details
 
     NETBIOS_NAMES = ["Name1", "Name2", "Name3"]
 
@@ -36,8 +36,8 @@ def test_get_dc_details_multiple_netbios_names(host, monkeypatch):
 
 
 @pytest.mark.slow
-def test_get_dc_details_no_netbios_names(host, monkeypatch):
-    from infection_monkey.exploit.zerologon_utils.vuln_assessment import get_dc_details
+def test_dc_utils__get_dc_details_no_netbios_names(host, monkeypatch):
+    from infection_monkey.exploit.zerologon_utils.dc_utils import get_dc_details
 
     NETBIOS_NAMES = []
